Handling cancel of file input
